adds back app.js and creates LSS
